requirements.txt
anthropic
notion-client
python-dotenv
httpx
flask
psycopg2-binary
werkzeug
apscheduler
